refactor: separate current schemas from migration contracts
