+++ docs/agents/architecture.md
@@ -33,12 +33,12 @@ Before changing background commands or sync execution, read [ADR-0001](../adr/00
 - Cards are slug-keyed, own their notes, and store dates as numbers.
 - Preserve supported card fields, FSRS schedules, notes, stats, settings, and Gist configuration. Unrecognized fields may be discarded; unrelated undecodable records need not survive mutations.
 - Schema changes require a new numbered module in `infrastructure/storage/migrations/`, appended to the sequence in `infrastructure/storage/migrations/runner.ts`.
-- Each migration declares its historical input and output contracts. A loader validates the preceding migration's output shape; its saver accepts its own output type. Shared schemas live in `domain/schemas/vN.ts`, named for the schema version that introduced them. Keep existing contracts fixed; reuse or extend them for later versions. Application models and migrations compose these definitions with their own normalization and extra-field handling. Migrations import a specific version, never current application models; historical storage keys remain independent of layout changes.
+- Each migration declares its historical input and output contracts. A loader validates the preceding migration's output shape; its saver accepts its own output type. Current schemas live in named domain modules. Migrations own frozen historical contracts and never import current application schemas; runtime callers never import migration contracts. Keep historical contracts and storage keys fixed, even where this duplicates current definitions.
 - Use `Input`, `Output`, `validateInput`, and `validateOutput` consistently within each migration. Export `Output` and `validateOutput` for the next migration, which imports them as `Input` and `validateInput`. When a local input validator adds checks or error context, import the preceding validator as `validatePreviousOutput` and call it from `validateInput`. For an identity migration, use `export type Output = Input` and alias the validator with `export const validateOutput: typeof validateInput = validateInput`. Validators accept `unknown` and assert their validated type; local helper types may express additional narrowing. Keep the lifecycle names `load`, `migrate`, `save`, and optional `cleanup`.
-- A physical layout reader knows where values are stored and how to assemble them into a raw dataset; it does not validate a migration's data schema. Each migration's loader validates those raw values against its historical input contract. Introduce a new layout reader when a migration needs to read changed storage locations or assembly rules; do not add unused readers.
+- A physical layout reader knows where values are stored and how to assemble them into a raw dataset; it does not validate a migration's data schema. Each migration's loader validates those raw values against its historical input contract. A migration that changes storage locations or assembly rules must supply and test the new layout reader.
 - Startup validates the completed version, then loads, transforms, saves, and cleans up each step before recording its version. The runner assumes no dataset shape. Backup parsing validates the declared version's contract before running the remaining pure transformations, including when none remain. Missing version means version 0.
 - Use best-effort retry safety: prepare and validate changes before writing, batch related writes to the same storage area where practical, and prefer operations that are safe to repeat, such as setting a value or removing an obsolete key. Record each migration's completed version only after its writes and cleanup succeed. Consider what loading and rerunning the step would encounter after an interruption; do not introduce recovery snapshots or a transaction framework without a concrete need.
-- Shared physical layout readers live in `infrastructure/storage/migrations/layouts/vN.ts` and export `readDataset`. `N` is the schema version that introduced the layout, not the only version it supports. Reuse a reader while the physical layout remains compatible. When a migration introduces a new layout, it reads the old layout and writes the new one. Retain the old reader; add a reader named for the version that introduced the new layout when a later migration needs it. Never rename an existing layout module to reflect which layout is current.
+- Shared physical layout readers live in `infrastructure/storage/migrations/layouts/vN.ts` and export `readDataset`. `N` is the schema version that introduced the layout, not the only version it supports. Reuse a reader while the physical layout remains compatible. When a migration introduces a new layout, retain the old reader and add one named for that migration’s version. The migration reads the old layout and writes the new one; its tests verify the result through the new reader so later migrations can use it. Never rename an existing layout module to reflect which layout is current.
 - Infrastructure owns card date codecs, stored-record validation, migrations, and persistence. Persistence adapters do not mark local edits.
 - Domain owns application schemas. Infrastructure owns the backup envelope, version and timestamp handling, field filtering, and backup-specific relationship checks through synchronous, storage-free `parseBackup(json): PreparedImportData`. It preserves raw historical fields until their migration runs, then applies current schemas and relationship checks. Forbidden historical fields are contract violations, not disposable extras.
 - Services prepare file imports and Gist pulls through `parseBackup` before replacement, then orchestrate reset/restore through shared persistence adapters. Preserve the local PAT and apply imported timestamps after settings writes.
