+++ domain/notes.ts
@@ -1,7 +1,8 @@
 import { z } from 'zod';
-import { noteTextSchema } from './schemas/v4';
-
-export { NOTES_MAX_LENGTH } from './schemas/v4';
+export const NOTES_MAX_LENGTH = 500;
+export const noteTextSchema = z.string().max(NOTES_MAX_LENGTH, {
+  error: `Note exceeds maximum length of ${NOTES_MAX_LENGTH} characters`,
+});
 
 export const noteSchema = z.object({
   text: noteTextSchema,
