--- domain/schemas/v5.ts
@@ -1,7 +1,0 @@
-import { z } from 'zod';
-
-export const gistConnectionSchema = z.object({
-  pat: z.string(),
-  gistId: z.string().nullable(),
-  enabled: z.boolean(),
-});
